requirements.txt
alembic==1.13.2
appdirs==1.4.4
attrs==24.2.0
bcrypt==3.1.7
blinker==1.8.2
certifi==2024.8.30
cffi==1.17.0
charset-normalizer==3.3.2
click==8.1.7
cryptography==2.8
Fabric3==1.14.post1
flasgger==0.9.7.1
Flask==3.0.3
Flask-Cors==5.0.0
flask-swagger-ui==4.11.1
greenlet==3.0.3
gunicorn==23.0.0
httplib2==0.22.0
idna==3.8
itsdangerous==2.2.0
Jinja2==3.1.4
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
Mako==1.3.5
MarkupSafe==2.1.5
mistune==3.0.2
mysqlclient==2.2.4
packaging==24.1
paramiko==2.12.0
psycopg2-binary==2.9.9
pycparser==2.22
PyNaCl==1.3.0
pyparsing==3.1.4
python-dotenv==1.0.1
PyYAML==6.0.2
redis==5.0.8
referencing==0.35.1
requests==2.32.3
rpds-py==0.20.0
setuptools==40.1.0
six==1.16.0
SQLAlchemy==2.0.32
typing_extensions==4.12.2
urllib3==2.2.2
Werkzeug==3.0.4
